requirements.txt
aiohttp>=3.8.0
python-dotenv>=0.19.0
asyncio-throttle>=1.0.0